refactor: 테스트시 필요한 요청 메소드를 유틸 클래스로 분리
