+++ src/test/java/wooteco/subway/line/SectionRestAssuredRequestUtils.java
@@ -1,0 +1,31 @@
+package wooteco.subway.line;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.MediaType;
+import wooteco.subway.line.dto.LineResponse;
+import wooteco.subway.line.dto.SectionRequest;
+import wooteco.subway.station.dto.StationResponse;
+
+public final class SectionRestAssuredRequestUtils {
+
+    private SectionRestAssuredRequestUtils() {
+    }
+
+    public static ExtractableResponse<Response> 지하철_구간_생성_요청(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
+        SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
+
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(sectionRequest)
+                .when().post("/lines/{lineId}/sections", line.getId())
+                .then().log().all()
+                .extract();
+    }
+
+    public static void 지하철_구간_등록되어_있음(LineResponse lineResponse, StationResponse upStation, StationResponse downStation, int distance) {
+        지하철_구간_생성_요청(lineResponse, upStation, downStation, distance);
+    }
+}
