+++ src/test/java/wooteco/subway/path/PathAcceptanceTest.java
@@ -23,9 +23,9 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 import static wooteco.subway.auth.AuthAcceptanceTest.로그인되어_있음;
 import static wooteco.subway.auth.AuthAcceptanceTest.회원_등록되어_있음;
-import static wooteco.subway.line.LineAcceptanceTest.지하철_노선_등록되어_있음;
-import static wooteco.subway.line.SectionAcceptanceTest.지하철_구간_등록되어_있음;
-import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
+import static wooteco.subway.line.LineRestAssuredRequestUtils.지하철_노선_등록되어_있음;
+import static wooteco.subway.line.SectionRestAssuredRequestUtils.지하철_구간_등록되어_있음;
+import static wooteco.subway.station.StationRestAssuredRequestUtils.지하철역_등록되어_있음;
 
 @DisplayName("지하철 경로 조회")
 public class PathAcceptanceTest extends AcceptanceTest {
@@ -38,7 +38,11 @@ public class PathAcceptanceTest extends AcceptanceTest {
     private StationResponse 남부터미널역;
 
     /**
-     * 교대역    --- *2호선* 10 ---   강남역 |                              | *3호선* 3                   *신분당선* 10 |                              | 남부터미널역  --- *3호선* 2 ---   양재
+     * 교대역  --- *2호선* 10 ---    강남역
+     * |                            |
+     * *3호선* 3                  *신분당선* 10
+     * |                            |
+     * 남부터미널역 --- *3호선* 2 ---  양재
      */
     @BeforeEach
     public void setUp() {
